chore: code re-factoring
