+++ credentials/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
     implementation project(':config')
+    implementation project(':storage')
 
     implementation 'org.slf4j:slf4j-api:2.0.9'
     implementation 'org.apache.httpcomponents.client5:httpclient5:5.4.3'
