+++ credentials/src/main/java/com/epam/aidial/core/credentials/data/credentials/ResourceSignInRequest.java
@@ -1,5 +1,7 @@
-package com.epam.aidial.core.config;
+package com.epam.aidial.core.credentials.data.credentials;
 
+import com.epam.aidial.core.config.AuthenticationType;
+import com.epam.aidial.core.config.CredentialsLevel;
 import com.fasterxml.jackson.annotation.JsonAlias;
 import lombok.Data;
