--- credentials/src/main/java/com/epam/aidial/core/credentials/exception/HttpStatus.java
@@ -1,66 +1,0 @@
-package com.epam.aidial.core.credentials.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum HttpStatus {
-
-    OK(200),
-    NOT_MODIFIED(304),
-    BAD_REQUEST(400),
-    UNAUTHORIZED(401),
-    FORBIDDEN(403),
-    NOT_FOUND(404),
-    METHOD_NOT_ALLOWED(405),
-    CONFLICT(409),
-    PRECONDITION_FAILED(412),
-    REQUEST_ENTITY_TOO_LARGE(413),
-    UNSUPPORTED_MEDIA_TYPE(415),
-    UNPROCESSABLE_ENTITY(422),
-    TOO_MANY_REQUESTS(429),
-    INTERNAL_SERVER_ERROR(500),
-    BAD_GATEWAY(502),
-    SERVICE_UNAVAILABLE(503),
-    GATEWAY_TIMEOUT(504),
-    HTTP_VERSION_NOT_SUPPORTED(505);
-
-    private final int code;
-
-    public boolean is5xx() {
-        return code >= 500 && code < 600;
-    }
-
-    public boolean is4xx() {
-        return code >= 400 && code < 500;
-    }
-
-    public static HttpStatus fromStatusCode(int code) {
-        return fromStatusCode(code, INTERNAL_SERVER_ERROR);
-    }
-
-    public static HttpStatus fromStatusCode(int code, HttpStatus fallback) {
-        return switch (code) {
-            case 200 -> OK;
-            case 304 -> NOT_MODIFIED;
-            case 400 -> BAD_REQUEST;
-            case 401 -> UNAUTHORIZED;
-            case 403 -> FORBIDDEN;
-            case 404 -> NOT_FOUND;
-            case 405 -> METHOD_NOT_ALLOWED;
-            case 409 -> CONFLICT;
-            case 412 -> PRECONDITION_FAILED;
-            case 413 -> REQUEST_ENTITY_TOO_LARGE;
-            case 415 -> UNSUPPORTED_MEDIA_TYPE;
-            case 422 -> UNPROCESSABLE_ENTITY;
-            case 429 -> TOO_MANY_REQUESTS;
-            case 500 -> INTERNAL_SERVER_ERROR;
-            case 502 -> BAD_GATEWAY;
-            case 503 -> SERVICE_UNAVAILABLE;
-            case 504 -> GATEWAY_TIMEOUT;
-            case 505 -> HTTP_VERSION_NOT_SUPPORTED;
-            default -> fallback;
-        };
-    }
-}
