--- credentials/src/main/java/com/epam/aidial/core/credentials/exception/ResourceNotFoundException.java
@@ -1,15 +1,0 @@
-package com.epam.aidial.core.credentials.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException() {
-    }
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-
-    public ResourceNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
