+++ credentials/src/main/java/com/epam/aidial/core/credentials/factory/NoneAuthResourceCredentialsFactory.java
@@ -1,17 +1,17 @@
-package com.epam.aidial.core.credentials.service.factory;
+package com.epam.aidial.core.credentials.factory;
 
 import com.epam.aidial.core.config.ResourceAuthSettings;
-import com.epam.aidial.core.config.ResourceSignInRequest;
 import com.epam.aidial.core.credentials.data.credentials.ResourceCredentials;
+import com.epam.aidial.core.credentials.data.credentials.ResourceSignInRequest;
 
-public class ApiKeyResourceCredentialsFactory implements ResourceCredentialsFactory {
+public class NoneAuthResourceCredentialsFactory implements ResourceCredentialsFactory {
 
     @Override
     public ResourceCredentials createCredentials(String resourceId,
                                                  ResourceAuthSettings authSettings,
                                                  ResourceSignInRequest signInRequest) {
-        if (signInRequest.getCode() != null) {
-            throw new IllegalArgumentException("Code is not required when auth type is API_KEY");
+        if (signInRequest.getApiKey() != null || signInRequest.getCode() != null) {
+            throw new IllegalArgumentException("Neither Api key nor Code is not required when auth type is None");
         }
 
         long currentTime = System.currentTimeMillis();
@@ -19,8 +19,6 @@ public class ApiKeyResourceCredentialsFactory implements ResourceCredentialsFact
             .resourceId(resourceId)
             .credentialsLevel(signInRequest.getCredentialsLevel())
             .authenticationType(signInRequest.getAuthenticationType())
-            .apiKeyHeader(authSettings.getApiKeyHeader())
-            .apiKey(signInRequest.getApiKey())
             .createdAt(currentTime)
             .updatedAt(currentTime)
             .build();
