+++ credentials/src/main/java/com/epam/aidial/core/credentials/service/ResourceCredentialsManager.java
@@ -3,14 +3,14 @@ package com.epam.aidial.core.credentials.service;
 import com.epam.aidial.core.config.AuthenticationType;
 import com.epam.aidial.core.config.CredentialsLevel;
 import com.epam.aidial.core.config.ResourceAuthSettings;
-import com.epam.aidial.core.config.ResourceSignInRequest;
-import com.epam.aidial.core.config.ResourceSignOutRequest;
 import com.epam.aidial.core.credentials.data.credentials.AuthorizationHeader;
 import com.epam.aidial.core.credentials.data.credentials.ResourceCredentials;
+import com.epam.aidial.core.credentials.data.credentials.ResourceSignInRequest;
+import com.epam.aidial.core.credentials.data.credentials.ResourceSignOutRequest;
 import com.epam.aidial.core.credentials.data.credentials.TokenResponse;
-import com.epam.aidial.core.credentials.exception.ResourceNotFoundException;
-import com.epam.aidial.core.credentials.service.factory.ResourceCredentialsFactory;
-import com.epam.aidial.core.credentials.service.factory.ResourceCredentialsFactoryProvider;
+import com.epam.aidial.core.credentials.factory.ResourceCredentialsFactory;
+import com.epam.aidial.core.credentials.factory.ResourceCredentialsFactoryProvider;
+import com.epam.aidial.core.storage.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
