+++ credentials/src/main/java/com/epam/aidial/core/credentials/service/ResourceRegistrationService.java
@@ -7,8 +7,8 @@ import com.epam.aidial.core.credentials.data.registration.ClientRegistration;
 import com.epam.aidial.core.credentials.data.registration.ClientRegistrationRequest;
 import com.epam.aidial.core.credentials.data.registration.ClientRegistrationResponse;
 import com.epam.aidial.core.credentials.exception.CredentialsInternalException;
-import com.epam.aidial.core.credentials.exception.HttpException;
-import com.epam.aidial.core.credentials.exception.HttpStatus;
+import com.epam.aidial.core.storage.http.HttpException;
+import com.epam.aidial.core.storage.http.HttpStatus;
 import com.nimbusds.oauth2.sdk.pkce.CodeChallengeMethod;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
