+++ server/src/main/java/com/epam/aidial/core/server/service/ToolSetService.java
@@ -1,15 +1,16 @@
 package com.epam.aidial.core.server.service;
 
 import com.epam.aidial.core.config.ToolSet;
+import com.epam.aidial.core.credentials.service.ResourceAuthSettingsService;
 import com.epam.aidial.core.server.data.ResourceTypes;
 import com.epam.aidial.core.server.util.ProxyUtil;
 import com.epam.aidial.core.storage.data.ResourceItemMetadata;
+import com.epam.aidial.core.storage.exception.ResourceNotFoundException;
 import com.epam.aidial.core.storage.resource.ResourceDescriptor;
 import com.epam.aidial.core.storage.service.ResourceService;
 import com.epam.aidial.core.storage.util.EtagHeader;
 import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.tuple.Pair;
-import com.epam.aidial.core.credentials.service.ResourceAuthSettingsService;
 
 @AllArgsConstructor
 public class ToolSetService {
